Refatorando retorno dos metodos da controller e adicionando captura de mensagens de validação

+++ Desafio 1/src/crud-pessoa.api/Controllers/PessoaController.cs
@@ -1,4 +1,6 @@
-﻿using crud_pessoa.api.Dtos;
+﻿using crud_pessoa.api.Configs.Notifications;
+using crud_pessoa.api.Dtos;
+using crud_pessoa.api.Dtos.Responses;
 using crud_pessoa.api.Services;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
@@ -12,10 +14,13 @@ namespace crud_pessoa.api.Controllers
     public class PessoaController : ControllerBase
     {
         private readonly IPessoaService _pessoaService;
+        private readonly INotificacaoContext _notificacaoContext;
 
-        public PessoaController(IPessoaService pessoaService)
+        public PessoaController(IPessoaService pessoaService,
+                                INotificacaoContext notificacaoContext)
         {
             _pessoaService = pessoaService;
+            _notificacaoContext = notificacaoContext;
         }
 
 
@@ -23,19 +28,30 @@ namespace crud_pessoa.api.Controllers
         [Route("listar-pessoa")]
         public async Task<ActionResult<IEnumerable<PessoaDto>>> GetPessoasAsync([FromQuery] Documento documento)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             var result = await _pessoaService.GetAllAsync(documento.Cpf);
 
             if (result == null)
-                return NotFound();
+                return BadRequest();
 
             return Ok(result);
         }
 
         [HttpPost]
         [Route("cadastrar-pessoa")]
-        public async Task<ActionResult<int?>> InsertPessoaAsync([FromBody] PessoaDto pessoaDto)
+        public async Task<ActionResult<ResultResponse>> InsertPessoaAsync([FromBody] PessoaDto pessoaDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             var result = await _pessoaService.InsertAsync(pessoaDto);
+
+            if (_notificacaoContext.HasNotification())
+            {
+                return BadRequest(_notificacaoContext.RetornarNoticacoes());
+            }
 
             if (result == null)
                 return NotFound();
@@ -45,25 +61,27 @@ namespace crud_pessoa.api.Controllers
 
         [HttpPut]
         [Route("atualizar-pessoa")]
-        public async Task<ActionResult<int?>> InsertPessoaAsync([FromBody] PessoaUpdateDto pessoaUpdateDto)
+        public async Task<ActionResult<ResultResponse>> InsertPessoaAsync([FromBody] PessoaUpdateDto pessoaUpdateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             var result = await _pessoaService.UpdateAsync(pessoaUpdateDto);
 
             if (result == null)
-                return NotFound();
+                return BadRequest();
 
             return Ok(result);
         }
 
         [HttpPut]
         [Route("deletar-pessoa/{id:int}")]
-        public async Task<ActionResult<int?>> DeletarPessoaAsync(int id)
+        public async Task<ActionResult<bool>> DeletarPessoaAsync(int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             var result = await _pessoaService.DeleteAsync(id);
-
-            if (result == null)
-                return NotFound();
-
             return Ok(result);
         }
     }
